MAJOR overhaul of source
Compiles EVERYTHING from source now. This repo is now only used to download
sources and maintain a clean build system. Hopefully at this point, it is
clean and attractive enough for a few more people to help out on this project.
I'll have to add a Documentation section for those new to this project so that
everyone is on the same page and knows how to submit patches. If you want to
help out, email me at rbheromax@gmail.com